Use partition 'ubi' (mtd12) for rootfs and rootfs_data; use 512 UBI sub-blocks
Note that if an old UBI partition exists with VID header offset 2048, then
the kernel will log errors when trying to attach this partition.

This is no problem if the system rootfs is in RAM (initramfs)
or if it is located on a non UBI partition.

When sysupgrade realizes that 'ubi' cannot be attached (due to old format),
it will ubiformat the partition for subpage usage und then install the rootfs.

Signed-off-by: arny <arnysch@gmx.net>