Fix LED activity level specification in .dts file
As Henning Schild already found out, the power LED is turned on (i.e.
"active") on high GPIO output level. Same is true for the TFT backlight.
(I think this is unusual; it is different from most other devices)

Up to now it was incorrectly specified in the .dts file, which resulted in
some complications. All software which wanted to turn on a LED needed to
write the nonstandard value 0 into the brightness (for the power LED) or the
bl_power (for the backlight) node in the sysfs.

Of course software controlling the LEDs need to be modified to write the
correct value (e.g. lcdcontroller package) into the sysfs node.

This fix also allows to remove the hack in file rc.local of package
"base-files", where so far the backlight LED is turned off.

Code has been successfully tried out (together with changes in package lcdcontroller).

Note that now on initializing the backlight, it will be turned on for half a second. I have not
searched for the reason why this happens, but I find this not to be a problem.

Signed-off-by: arny arnysch@gmx.net